1. Have added a sleuth of filesystem commands  a) filesystem_write  b) filesystem_append  c) filesystem_readall  d) filesystem_readtolist  e) filesystem_delete  f) filesystem_copy  g) filesystem_move  h) filesystem_sethidden  i) filesystem_setvisible  j) filesystem_mkdir  k) filesystem_rmdir  l) filesystem_copydir  m) filesystem_movedir 2. Add [List:Length:name] to SetVarVal and print (allowing users to get number of items in a list) 3. Added [Filesystem:Fileexists:path] and [Filesystem:Direxists:path] which returns True or False if the file/directory exists (will make it easier to make a bool specific to existence of a file/directory)
Most filesystem commands work in typical CLI mode. I have come across a very unique bug while trying to make a primitive text editor with a 'Save' button - a string takes the textview's text, but is unable to output the value into a file via filesystem_write or filesystem_append. I can not find any documentation on this. The string was able to make the console title equal the textview's text. Was stuck on this for about 2 hours.

* do some basic network stuff (ping, TCP, etc)