1. Have added 'msgbox' with OK and YESNO. Workaround previous issue with threading via MainLoop.Invoke()=> 2. Changed [\n] to [nl] because [\n] was resulting in a compilation error when .gs scripts were compiled into .exe binaries. Simple fix.
Known issue: Invoking a messagebox into the main loop that returns an int for pressed button is not being captured outside the invocation. Movinf the if/then statement inside the invocation doesn't work either. Currently stuck on how to pass the 1 or 0 from the YESNO messagebox to the cooresponding integer, so it can change the assigned bool's value accordingly.