1. Added checkbox GUI items  a) Checkboxes can be linked to bool variables upon initalization 2. Some other general debugging
Note: Background tasks are likely just going to be "officially not recommended" when dealing with GUI components. UI is not thread safe to begin with. Safeguards can be put in so it's better managed but for now on less attention is going to be spent trying to deal with this and redirected to expanding and broadening functionality.