Preprocessor work -
A lot of preprocessor work has been implemented in order to help shave the end result file size of A) the GyroPrompt interpreter, and B) compiled .exe binaries from the GyroScript IDE (which was including a lot of unused code and bloat which did not always need to be included)

This is still work-in-progress. Main goal it trimming down the size of .exe's the IDE outputs when compiling .gs scripts.