Compiler optimizations ~
Most of these updates are related to assisting the compiler trim the size of the output binary when compiled in the GyroScript IDE. It's hard to tell when the IDE source isn't posted up (this will eventually change).

This was mostly about preprocessor statements being placed through out the code. While this makes it very unpleasant to work with, it will be necessary to keep preprocessor configuration up-to-date in tandem with feature rollout.

While the preprocessor statements are the top of Parser.cs are not global (despite references to them in other classes) that is because the constants are declared in the .csproj file as custom symboles (they are in Parser.cs to allow for easier debugging)